update snippets and order
