+++ .snippets/code/tutorials/by-product/wormholescan/replace-signatures/replace-sigs-2.ts
@@ -5,6 +5,9 @@ export const ETH_CORE =
 
 export const WORMHOLE_SCAN_API = 'https://api.wormholescan.io/v1';
 
+export const LOG_MESSAGE_PUBLISHED_TOPIC =
+  '0x6eb224fb001ed210e379b335e35efe88672a8ce935d981a6896b27ffdf52a3b2';
+
 export const TXS = [
   '0x3ad91ec530187bb2ce3b394d587878cd1e9e037a97e51fbc34af89b2e0719367',
   '0x3c989a6bb40dcd4719453fbe7bbac420f23962c900ae75793124fc9cc614368c',
