+++ .snippets/code/tutorials/by-product/wormholescan/replace-signatures/replace-sigs-5.ts
@@ -1,18 +1,12 @@
-import { fetchVaaId, fetchVaa } from '../src/helpers/vaaHelper';
+import { fetchVaaId } from '../src/helpers/vaaHelper';
 import { TXS } from '../src/config/constants';
 
-const testFetchVaa = async () => {
+const testFetchVaaId = async () => {
   const vaaId = await fetchVaaId(TXS[0]);
 
-  if (!vaaId) {
-    console.log('VAA ID not found.');
-    return;
+  if (vaaId) {
+    console.log(`VAA ID: ${vaaId}`);
   }
-
-  const vaaBytes = await fetchVaa(vaaId);
-  console.log(
-    `VAA Bytes: ${vaaBytes ? vaaBytes.toString('hex') : 'Not found'}`
-  );
 };
 
-testFetchVaa();
+testFetchVaaId();
