+++ .snippets/code/tutorials/by-product/wormholescan/replace-signatures/replace-sigs-6.ts
@@ -1,25 +1,18 @@
-import {
-  fetchVaaId,
-  fetchVaa,
-  checkVaaValidity,
-} from '../src/helpers/vaaHelper';
+import { fetchVaaId, fetchVaa } from '../src/helpers/vaaHelper';
 import { TXS } from '../src/config/constants';
 
-const testCheckVaaValidity = async () => {
+const testFetchVaa = async () => {
   const vaaId = await fetchVaaId(TXS[0]);
+
   if (!vaaId) {
     console.log('VAA ID not found.');
     return;
   }
 
   const vaaBytes = await fetchVaa(vaaId);
-  if (!vaaBytes) {
-    console.log('VAA not found.');
-    return;
-  }
-
-  const result = await checkVaaValidity(vaaBytes.toString('base64'));
-  console.log('VAA Validity:', result);
+  console.log(
+    `VAA Bytes: ${vaaBytes ? vaaBytes.toString('hex') : 'Not found'}`
+  );
 };
 
-testCheckVaaValidity();
+testFetchVaa();
